pyproject.toml
[tool.poetry]
name = "fastapi-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Alangard <73427768+Alangard@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
alembic = "^1.13.2"
uvicorn = "^0.30.1"
pydantic = "^2.7.4"
SQLAlchemy = "^2.0.31"
websocket = "^0.2.1"
pydantic-settings = "^2.3.4"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
aiogram = "^3.8.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
